requirements.txt
GitPython==3.1.41
cerberus==1.3.5
PyYAML==6.0.1

